moves sorted tests to OperatorToObservableSortedListTest
